Model.load, adding checkpoints
Propose to add model.fit under the function of load, as it enables user to run model.load in their script, and to also run the preprocessor and save all images into the disk, and train based on the previous best model. This combats the issue of some users having limited GPU due to factors such as Google Colab, or even due to power outages and a disruption in Internet connection.

I am unsure if the fork that I am proposing is accurate, however it is due to evaluation by @muellerdo and team.

Thank you so much for the attention, and I wish you the best of health. 

Best regards